Finished most geoSearch functionality
Geocoder now locates points within a certain distance, displays them on
map and populates the sidebar. Needed to restructure the data -> geoJson
process and put the 'pointToLayer' and 'onEachFeature' stuff into
functions so they can be re-used after geo searches, and eventually with
the float plan tool.

Still some stylistic aspects of this need to be accomplished.