used new co-id system to network confluences
Still an issue with arranging the functions correctly. When using a
while loop, it runs through the variables in the loop first, which
produces an undefined when it tries to reference the index value of the
array of stream mile objects. Have to figure out how to arrange it all
properly, but at least now I know.