added float results modal
The float results modal will replace the float plan modal when 'generate
plan' is clicked. a varaible is set to 'true' that sets the state of the
float plan. if it is true, the float plan must be cleared to make it
false and re-initiate the float plan buttons in feature modal.

results modal reports stats from the float plan - total mileage and
number of rivers... more to come?